✨ feat: Add black-magic.png and update README.md
This commit adds a new file, black-magic.png, to the assets directory. Additionally, the README.md file has been updated with a mention of makefile blackmagic and an embedded image of black-magic.png.

Note: The commit message is based on the provided diff and assumes that these changes are intended to be committed as a new feature (`feat`) in accordance with conventional commits specification.